Convert the list of numeric values to a one-dimensional NumPy array.

+++ array01.py
@@ -5,5 +5,5 @@ def main(array):
     """
     Convert the list of numeric values to a one-dimensional NumPy array.
     """
-    a = np.array(array, dtype=np.float8)
+    a = np.array(array)
     return a
